Polish implementation
* Polish implementation
* Added documentation for service
* Adding descriptive documentation
Ticket: SYM-80, SYM-89
PRs: #16
