+++ Sources/ForageSDK/Foundation/Network/LiveForageService.swift
@@ -4,21 +4,25 @@
 //
 //  Created by Symphony on 30/10/22.
 //
+
 import Foundation
 import VGSCollectSDK
 
 internal class LiveForageService: ForageService {
-
+    
     // MARK: Properties
     
     internal var provider: Provider
     private var collector: VGSCollect?
-
+    private var maxAttempts: Int = 10
+    private var intervalBetweenAttempts: Double = 1.0
+    private var retryCount = 0
+    
     init(_ collector: VGSCollect?, provider: Provider = Provider()) {
         self.collector = collector
         self.provider = provider
     }
-
+    
     // MARK: Tokenize EBT card
     
     /// Tokenize a given *ForagePANRequestModel* object
@@ -30,7 +34,7 @@ internal class LiveForageService: ForageService {
         do { try provider.execute(endpoint: ForageAPI.tokenizeNumber(request: request), completion: completion) }
         catch { completion(.failure(error)) }
     }
-
+    
     // MARK: X-key
     
     /// Retrieve from ForageAPI the X-Key header to perform request.
@@ -42,12 +46,12 @@ internal class LiveForageService: ForageService {
         do { try provider.execute(model: ForageXKeyModel.self, endpoint: ForageAPI.xKey(bearerToken: bearerToken), completion: completion) }
         catch { completion(.failure(error)) }
     }
-
+    
     // MARK: Cancel request
     
     /// Cancel any ongoing request.
     internal func cancelRequest() { provider.stopRequestOnGoing() }
-
+    
     // MARK: Check balance
     
     /// Perform VGS SDK request to retrieve balance.
@@ -71,33 +75,35 @@ internal class LiveForageService: ForageService {
 
         collector?.sendData(
             path: "/api/payment_methods/\(request.paymentMethodReference)/balance/",
-            extraData: extraData) { result in
-                switch result {
-                case .success(_, let data, _):
-                    completion(.success(data))
-                case .failure(let code, let data, let response, let error):
-                    if let data = data {
-                        self.provider.processVGSData(
-                            model: ForageServiceError.self,
-                            code: code,
-                            data: data,
-                            response: response) { errorResult in
-                            switch errorResult {
-                            case .success(let errorParsed):
-                                return completion(.failure(errorParsed))
-                            case .failure(let error):
-                                return completion(.failure(error))
-                            }
-                        }
-                    } else if let error = error {
-                        return completion(.failure(error))
-                    } else {
-                        return completion(.failure(ServiceError.emptyError))
+            extraData: extraData) { [weak self] result in
+                self?.polling(response: result, request: request, completion: { pollingResult in
+                    switch pollingResult {
+                    case .success:
+                        self?.retrieveCheckBalance(
+                            request: request,
+                            completion: completion
+                        )
+                    case .failure(let error):
+                        completion(.failure(error))
                     }
-                }
+                })
             }
     }
-
+    
+    /// Perform request to Forage API after polling got message success.
+    ///
+    /// - Parameters:
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Returns balance object.
+    internal func retrieveCheckBalance(request: ForageRequestModel,
+                                       completion: @escaping (Result<Data?, Error>) -> Void) {
+        do {
+            try provider.execute(endpoint: ForageAPI.retrieveBalance(request: request), completion: completion)
+        } catch {
+            completion(.failure(error))
+        }
+    }
+    
     // MARK: Capture payment
     
     /// Perform VGS SDK request to capture payment.
@@ -122,29 +128,127 @@ internal class LiveForageService: ForageService {
         collector?.sendData(
             path: "/api/payments/\(request.paymentReference)/capture/",
             extraData: extraData) { result in
-                switch result {
-                case .success(_, let data, _):
-                    return completion(.success(data))
-                case .failure(let code, let data, let response, let error):
-                    if let data = data {
-                        self.provider.processVGSData(
-                            model: ForageServiceError.self,
-                            code: code,
-                            data: data,
-                            response: response) { errorResult in
-                            switch errorResult {
-                            case .success(let errorParsed):
-                                return completion(.failure(errorParsed))
+                self.polling(response: result, request: request, completion: { pollingResult in
+                    switch pollingResult {
+                    case .success:
+                        self.retrieveCapturedPayment(
+                            request: request,
+                            completion: completion
+                        )
+                    case .failure(let error):
+                        completion(.failure(error))
+                    }
+                })
+            }
+    }
+    
+    /// Perform request to Forage API after polling got message success.
+    ///
+    /// - Parameters:
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Returns captured payment object.
+    internal func retrieveCapturedPayment(
+        request: ForageRequestModel,
+        completion: @escaping (Result<Data?, Error>) -> Void)
+    {
+        do { try provider.execute(endpoint: ForageAPI.retrieveCapturedPayment(request: request), completion: completion) }
+        catch { completion(.failure(error)) }
+    }
+}
+
+// MARK: - Polling
+
+extension LiveForageService: Polling {
+    /// Process VGSData for polling message.
+    ///
+    /// - Parameters:
+    ///  - response: The *VGSResponse* which contains the result from VGS SDK.
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Which will return the result.
+    internal func polling(response: VGSResponse, request: ForageRequestModel, completion: @escaping (Result<Data?, Error>) -> Void) {
+        retryCount = 0
+        
+        switch response {
+        case .success(_, let data, _):
+            provider.processVGSData(model: MessageResponseModel.self, code: nil, data: data, response: nil) { [weak self] messageResponse in
+                switch messageResponse {
+                case .success(let message):
+                    self?.pollingMessage(
+                        message: message,
+                        request: request) { pollingResult in
+                            switch pollingResult {
+                            case .success:
+                                completion(.success(nil))
                             case .failure(let error):
-                                return completion(.failure(error))
+                                completion(.failure(error))
                             }
                         }
-                    } else if let error = error {
-                        return completion(.failure(error))
-                    } else {
-                        return completion(.failure(ServiceError.emptyError))
-                    }
+                case .failure(let error):
+                    completion(.failure(error))
                 }
             }
+
+        case .failure(_, _, _, let error):
+            guard let error = error else {
+                return completion(.failure(ServiceError.emptyError))
+            }
+            completion(.failure(error))
+        }
+    }
+    
+    /// Polls message to check payment status
+    ///
+    /// - Parameters:
+    ///  - message: The *MessageResponseModel* which contains the message.
+    ///  - request: Model element with data to perform request.
+    ///  - completion: Which will return the message for another retry or success.
+    internal func pollingMessage(
+        message: MessageResponseModel,
+        request: ForageRequestModel,
+        completion: @escaping (Result<MessageResponseModel, Error>) -> Void) -> Void
+    {
+        do {
+            try provider.execute(model: MessageResponseModel.self, endpoint: ForageAPI.message(request: message, bearerToken: request.authorization, merchantID: request.merchantID), completion: { [weak self] result in
+                guard let self = self else { return }
+                switch result {
+                case .success(let data):
+                    /// check message is not failed and is completed
+                    if data.failed == false && data.status == .completed {
+                        completion(.success(data))
+                    /// check message is failed to return error immediately
+                    } else if data.failed == true {
+                        completion(.failure(NSError(domain: "Message failed", code: 001, userInfo: nil)))
+                    /// check maxAttempts to retry
+                    } else if self.retryCount < self.maxAttempts {
+                        self.waitNextAttempt {
+                            self.pollingMessage(
+                                message: data,
+                                request: request,
+                                completion: completion
+                            )
+                        }
+                    /// in case run out of attempts
+                    } else {
+                        completion(.failure(NSError(domain: "Reached max attempts", code: 429, userInfo: nil)))
+                    }
+                    
+                case .failure(let error):
+                    completion(.failure(error))
+                }
+            })
+        } catch {
+            completion(.failure(error))
+        }
+    }
+    
+    /// Support function to update retry count and interval between attempts.
+    ///
+    /// - Parameters:
+    ///  - completion: Which will return after an wait.
+    private func waitNextAttempt(completion: @escaping () -> ()) {
+        retryCount = retryCount + 1
+        DispatchQueue.main.asyncAfter(deadline: .now() + self.intervalBetweenAttempts) {
+            completion()
+        }
     }
 }
